refactor: move providers to app/_providers
