+++ www/admin/app/dashboard/_components/Header.tsx
@@ -14,7 +14,7 @@ import {
 	Text,
 } from "@mantine/core";
 import { useMenuStore } from "@admin/src/data/stores/dashboard";
-import { poppins } from "@admin/app/mantine";
+import { poppins } from "@admin/app/_providers/mantine";
 import { getCardColor } from "@admin/src/utils/colors";
 import UserMenu from "./UserMenu";
 import type { User } from "api/prisma/types";
